money-backend/requirements.txt
alembic==1.15.2
babel==2.17.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
cryptography==44.0.2
ecdsa==0.19.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==21.2.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pathvalidate==3.2.3
pycparser==2.22
Pygments==2.19.1
PyMySQL==1.1.0
python-dotenv==1.0.1
requests==2.32.3
requests-futures==1.0.2
shtab==1.7.2
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
